fix: moved file and deleted unnecessary
